--- a/pandas/core/groupby.py
+++ b/pandas/core/groupby.py
@@ -2193,9 +2193,9 @@ def _convert_grouper(axis, grouper):
         return grouper.get
     elif isinstance(grouper, Series):
         if grouper.index.equals(axis):
-            return grouper.values
+            return grouper._values
         else:
-            return grouper.reindex(axis).values
+            return grouper.reindex(axis)._values
     elif isinstance(grouper, (list, Series, Index, np.ndarray)):
         if len(grouper) != len(axis):
             raise AssertionError('Grouper and axis must be same length')