--- a/pandas/sparse/scipy_sparse.py
+++ b/pandas/sparse/scipy_sparse.py
@@ -30,7 +30,7 @@ def _to_ijv(ss, row_levels=(0,), column_levels=(1,), sort_labels=False):
     _check_is_partition([row_levels, column_levels], range(ss.index.nlevels))
 
     # from the SparseSeries: get the labels and data for non-null entries
-    values = ss._data.values._valid_sp_values
+    values = ss._data.internal_values()._valid_sp_values
 
     nonnull_labels = ss.dropna()