… asv_bench/asv.conf.json
… asv_bench/benchmarks/binary_ops.py
… doc/source/basics.rst
… doc/source/release.rst
… doc/source/timeseries.rst
… doc/source/whatsnew/v0.17.0.txt
… pandas/core/common.py
… pandas/core/index.py
… pandas/core/indexing.py
… pandas/core/internals.py
… pandas/core/ops.py
… pandas/core/series.py
… pandas/io/pytables.py
… pandas/io/tests/data/legacy_hdf/datetimetz_object.h5
… pandas/io/tests/data/legacy_msgpack/0.16.2/0.16.2_x86_64_darwin_2.7.10.msgpack
… pandas/io/tests/data/legacy_msgpack/0.16.2/0.16.2_x86_64_darwin_3.4.3.msgpack
… pandas/io/tests/data/legacy_pickle/0.16.2/0.16.2_x86_64_darwin_2.7.10.pickle
… pandas/io/tests/data/legacy_pickle/0.16.2/0.16.2_x86_64_darwin_3.4.3.pickle
… pandas/io/tests/generate_legacy_storage_files.py
… pandas/io/tests/test_packers.py
… pandas/io/tests/test_pickle.py
… pandas/io/tests/test_pytables.py
… pandas/io/tests/test_sql.py
… pandas/tests/test_base.py
… pandas/tests/test_categorical.py
… pandas/tests/test_dtypes.py
… pandas/tests/test_frame.py
… pandas/tests/test_index.py
… pandas/tests/test_internals.py
… pandas/tests/test_multilevel.py