@jreback 2015-07-01
This is the latest https://github.com/jreback/pandas/commit/19ec61d3bcbe84dada3f38bad3daa061a78707af I think

```
In [3]: from_dict
Out[3]: 
           a
0 2014-01-01
1 2014-01-02
2 2014-01-03
3 2014-01-04
4 2014-01-05
5 2014-01-06
6 2014-01-07
7 2014-01-08
8 2014-01-09
9 2014-01-10

In [4]: from_dict.dtypes
Out[4]: 
a    datetime64[ns, UTC]
dtype: object
```

``Series.values`` IS a ``DatetimeIndex`` that is how its implemented. Its very similar to how ``Sparse/Categorical`` are done. This preserves the tz info inside the object (actually have freq too). Rather than relying upon a ndarray impl and passing it around. Much cleaner this way.