@ssanderson 2015-07-01
I pulled 2422fe50fcc5aac541a0be8d67a4e86309b3e2ec ~20 minutes ago.  The commit hash would change if the content had changed.