@jreback 2015-07-01
did you make? (it has a tad bit of cython code changes)