@ssanderson 2015-07-01
I did a `pip install -e .` in a fresh virtualenv.  What in the above that I posted is different from what you're seeing?