@jreback 2015-07-01
sorry, I didn't see what you meant above. ``DataFrame(new_index)`` hang on a sec. Didnt have a test for that.