@ssanderson 2015-07-01
The from_dict case is working as expected modulo the unexpected type of `.values`, which sounds like it's actually as-designed.  The one that I think is incorrect is direct construction from a DatetimeIndex.