@jreback 2015-07-02
ok, indexing is fixed. wasn't testing things fully :mask: